+++ Microservices/Test_Game_Server/TestServerMain.cs
@@ -16,7 +16,7 @@ namespace Test_Game_Server
         static void Main(string[] args)
         {
             CommonLibrary.Parser.ParseCommandLine(args);
-            int applicationId = CommonLibrary.Parser.ApplicationId;
+            Int64 applicationId = CommonLibrary.Parser.ApplicationId;
             if (applicationId == 0)
             {
                 applicationId = Network.Utils.GetIPBasedApplicationId();
